Extract filtering of matches into own class
